[Media notas] -> Correção ToSingle()

+++ Prompt program/Media_notas.cs
@@ -5,13 +5,13 @@ namespace Media{
 			float som = 0, med;
 			string frase = "Digite a {0}º nota: ", situacao = "";
 			Console.WriteLine(frase, 1);
-			som += Convert.ToInt32(Console.ReadLine());
+			som += Convert.ToSingle(Console.ReadLine());
 			Console.WriteLine(frase, 2);
-			som += Convert.ToInt32(Console.ReadLine());
+			som += Convert.ToSingle(Console.ReadLine());
 			Console.WriteLine(frase, 3);
-			som += Convert.ToInt32(Console.ReadLine());
+			som += Convert.ToSingle(Console.ReadLine());
 			Console.WriteLine(frase, 4);
-			som += Convert.ToInt32(Console.ReadLine());
+			som += Convert.ToSingle(Console.ReadLine());
 			med = som/4;
 			Console.WriteLine("\nA media final foi......: " + med);
 			if(med>=7){
